requirements.txt
anyio==3.6.1
charset-normalizer==2.1.0
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
numpy==1.23.1
openai==0.20.0
openpyxl==3.0.10
pandas==1.4.3
pandas-stubs==1.4.3.220710
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
